Modularize the TypeForm question IDs.
The TypeForm IDs will also be referenced in testing so moving them into a module
means that there won't be opaque IDs in various places. The intention with the
question Strings is that they're helpful for development -- they don't need to
be an exact match with the wording to users or anything.