Use PostgreSQL as the test database.
PostgreSQL is needed instead of SQLite because of the use of the `hstore`
datatype for User#signup_blob. Unfortunately this means that to run local tests
locally, you'll need a running PostgreSQL database.

On OSX this can be done by:

Installing PostgreSQL:

- Run `brew install postgresql`

- Run: `createdb`

This sets up a database. By default it'll be named after your username.

After that, to start PostgreSQL you can:

- Run `pg_ctl -D /usr/local/var/postgres start`

And then to stop PostgreSQL you can:

- Run `pg_ctl -D /usr/local/var/postgres stop`

It's a bit more intense than ideal and perhaps this is incentive enough against
using `hstore` in place of just dumping JSON and re-parsing it if necessary.